+++ android/SessionLedger/mobile/src/main/java/press/pelldom/sessionledger/mobile/AppConfig.kt
@@ -1,0 +1,5 @@
+package press.pelldom.sessionledger.mobile
+
+object AppConfig {
+    const val ADS_ENABLED = false
+}